drivers: melfas_mss100: Ensure 'virtual touchpad' is freed when unloading driver
Missed a couple of cleanups following ea3e486, where 'virtual touchpad' support for Samsung DeX is now enabled by default regardless of DT requesting for the feature. Remove all references to 'support_dex' and ensure the 'virtual touchpad' it creates is properly destroyed and cleaned up when unloading the driver.

Signed-off-by: John Vincent <git@tenseventyseven.cf>